backend/pyproject.toml
[tool.poetry]
name = "intuitive-care-backend"
version = "0.1.0"
description = "Backend - Teste de entrada Intuitive Care"
authors = ["Pedro Garcia <pedrofsgarcia.pro@gmail.com>"]
packages = [{include = "api"}, {include = "etl"}, {include = "database"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.5"
pandas = "^3.0.0"
pydantic = "^2.12.5"
types-requests = "^2.32.4.20260107"
openpyxl = "^3.1.5"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.14"
mypy = "^1.19.1"
taskipy = "^1.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["venv", "__pycache__"]

[tool.ruff.lint]
select = [
    "E",  # Pycodestyle errors
    "W",  # Pycodestyle warnings
    "F",  # Pyflakes
    "I",  # Import order
    "N",  # Naming conventions
    "B",  # Bugbear (boas práticas e possíveis bugs)
    "C4", # Comprehensions
    "SIM",# Simplifiable code
    "UP", # Upgrades de código Python (melhorias automáticas)
]
ignore = [
    "F822",  # undefined name in __all__
    "F704",  # unused awaitable
]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["pandas.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.bandit]
exclude_dirs = [".venv", "venv"]
skips = ["B101"]  # Permite uso de assert

[tool.taskipy.tasks]
parte_um = "python -m etl.download"
consolidate = "python -m etl.run_ex_1"

# Parte 2: Transformação e validação
aggregate = "python -m etl.run_ex_2"

# Banco de dados
init_db = "python -m database.init_db"

api = "python -m api.api"

# Pipeline completo ETL (Partes 1-2)
etl = "task download && task consolidate && task transform"
